New Account Generation Page
The New Account component (should) contain all of the necessary UI to instantiate an account. Doesn't have any event handlers that actually send info to Firebase yet, but it's pretty slick. I also set it as the default on the home page when you log in but this can obviously be changed.